requirements.txt
numpy
scipy
biopython